Avoid adding trailing slash
I think we should let the user do this explicitly instead of having it
done implicitly because it can cause weird behaviour. For example, this
causes a double slash to occur when `$request_uri` is appended to it ...

```
return ${REDIRECT_CODE} ${REDIRECT_TARGET}\$request_uri;
```

This is because `$request_uri` will always have a `/` at the beginning.